Move neutron Python tests to match module hierarchy
